requirements.txt
Django==2.2.3
djangorestframework==3.9.2
pipfile==0.0.2
